+++ include/vlog/segment.h
@@ -44,7 +44,7 @@ struct SegmentSorter {
     bool operator ()(const size_t f1, const size_t f2) const {
         assert(f1 < maxSize);
         assert(f2 < maxSize);
-        for (uint8_t i = 0; i < vectors.size(); ++i) {
+        for (int i = 0; i < vectors.size(); ++i) {
             if ((*vectors[i])[f1] != (*vectors[i])[f2])
                 return (*vectors[i])[f1] < (*vectors[i])[f2];
         }
@@ -219,7 +219,7 @@ class Segment {
 #endif
 
         size_t getNRows() const {
-            for (uint8_t i = 0; i < nfields; ++i) {
+            for (int i = 0; i < nfields; ++i) {
                 if (columns[i] != NULL) {
                     if (!columns[i]->isEDB()) {
                         return columns[i]->size();
@@ -228,7 +228,7 @@ class Segment {
             }
 
             //What to do now?
-            for (uint8_t i = 0; i < nfields; ++i) {
+            for (int i = 0; i < nfields; ++i) {
                 if (columns[i] != NULL) {
                     return columns[i]->size();
                 }
@@ -237,7 +237,7 @@ class Segment {
         }
 
         bool isEmpty() const {
-            for (uint8_t i = 0; i < nfields; ++i) {
+            for (int i = 0; i < nfields; ++i) {
                 if (columns[i] != NULL) {
                     if (!columns[i]->isEmpty())
                         return false;
@@ -297,7 +297,7 @@ class Segment {
 
         uint8_t getNConstantFields() const {
             uint8_t n = 0;
-            for (uint8_t i = 0; i < nfields; ++i) {
+            for (int i = 0; i < nfields; ++i) {
                 if (columns[i] == NULL || columns[i]->isConstant()) {
                     n++;
                 }
@@ -314,7 +314,7 @@ class Segment {
         }
 
         bool isEDB() const {
-            for (uint8_t i = 0; i < nfields; ++i) {
+            for (int i = 0; i < nfields; ++i) {
                 if (columns[i] != NULL) {
                     if (!(columns[i]->isEDB())) {
                         return false;
